% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_funcs.R
\name{age}
\alias{age}
\title{Age in years}
\usage{
age(from, to = strptime(lubridate::now(), format = "\%Y-\%m-\%d"))
}
\arguments{
\item{from}{date POSIXlt/string or similar to be converted to POSIXlt.
e.g. "YYYY-MM-DD"}

\item{to}{date POSIXlt/string or similar to be converted to POSIXlt.
Defaults to to now()}
}
\description{
Calculates age in years
}
\examples{
age("1988-12-19", "2017-11-22")
# 28
}
\keyword{age}
