% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{decade}
\alias{decade}
\title{Get decade of a date-time object}
\usage{
decade(x)
}
\arguments{
\item{x}{a date-time object, e.g. Date, POSIXct, POSIXlt and others.}
}
\description{
Returns the "decade" of a date-time object as an integer
representing the first year of the decade. e.g. 1950L for the 1950's
}
\details{
Note: Setting decade is not supported.
}
\examples{
decade(as_date("1955-01-01"))
# 1950
}
\keyword{date-time}
\keyword{decade}
