% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grotta_bars.R
\name{grotta_bars}
\alias{grotta_bars}
\title{"Grotta" bars for comparing modified Rankin Scale}
\usage{
grotta_bars(data, mrs, bar_group, bar_width = 0.5)
}
\arguments{
\item{data}{data.frame with columns for `mrs` and `group`}

\item{mrs}{character string indicating variable in `data` for mRS. mRS data
should be numeric or integers values in (0,1,2,3,4,5,6)}

\item{bar_group}{character string indicating variable in `data` to be used for bars}

\item{bar_width}{widths of bars as a proportion, numeric between 0 and 1}
}
\description{
Charts comparing modified Rankin Scale of two groups using a stacked bar for each
group, which a diagonal linking equivalent levels are known as "Grotta" bars.
}
\examples{

# simple example
pl <- grotta_bars(data = mRS,
                 mrs = "mRS",
                 bar_group = "intv")

# ggplot object can be further adjusted as usual
# theme: note that some themes will move the legend
library(ggplot2)
pl +
 theme_bw()

# title, fill/colour palette, facets
pl +
 labs(title = "mRS pre- vs post-intervention") +
 scale_fill_brewer(palette = "Blues") +
 facet_grid(rows = vars(group))

# more complicated example, for a very simple chart
#change data, adding the variable bar_n
\dontrun{
pl2 <- pl

pl2$data <- pl2$data \%>\%
 dplyr::group_by(intv) \%>\%
 dplyr::mutate(bar_n = dplyr::n()) \%>\%
 dplyr::ungroup()

# access bar count data with after_stat(count). In this case we weight by 1/bar_n
# (ie divide by total count) and multiply by 100 to get within bar percentages.
# label uses ifelse() and sprintf to prevent printing in narrow bars and to set a
# sensible number of decimal places

pl2 +
 # percentage labels
 geom_text(aes(label = ifelse(after_stat(count) < 0.05,
                              "",
                              sprintf("\%0.0f\%\%", 100 * after_stat(count))),
               x = after_stat(count),
               weight = 1/bar_n),
           stat = "count",
           position = position_fill(vjust = 0.5)) +
 # similar idea for mRS labels
 geom_text(aes(label = mRS,
               x = after_stat(count),
               y = stage(start = intv,
                         after_scale = y + (y - 1.5)*0.6),
               weight = 1/bar_n),
           size = 4,
           stat = "count",
           position = position_fill(vjust = 0.5)) +
 # change colour scale
 scale_fill_brewer(palette = "Blues") +
 # change theme elements
 theme_void() +
 theme(axis.text.y = element_text()) +
 #remove legend
 guides(fill = "none") +
 # label bars ie y axis
 scale_y_discrete(labels = c(pre = "Pre-intervention", post = "Post-intervention"))}



}
\keyword{"grotta}
\keyword{bars"}
\keyword{grotta}
\keyword{mRS}
