% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{is wholenumer?}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric object to be test}

\item{tol}{tolerence, default is sq root of machine epsilon}
}
\description{
Tests if numeric object is a whole number within a given tolerence.
Note: I copied this function from the web and I don't remember where from.
}
\examples{
is.wholenumber(1.00000001, tol = 1.5e-08)
# TRUE
is.wholenumber(1.0000001, tol = 1.5e-08)
# FALSE
}
