% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_liners.R
\name{lmprint}
\alias{lmprint}
\title{Print a loading matrix}
\usage{
lmprint(loadings, cutoff = 0.35, digits = 3)
}
\arguments{
\item{loadings}{numerical matrix of variable loadings e.g.
from princomp()$loadings}

\item{cutoff}{numbers <= to cutoff will not be printed}

\item{digits}{number of significant digits to print}
}
\description{
Prints a matrix with all entries below a given cut off removed. Useful
to print loading matrices in factor analysis.
}
\examples{
set.seed(10)
lmprint(matrix(runif(9), 3, 3))
}
\keyword{fa}
\keyword{factanal}
\keyword{pca}
\keyword{princomp}
