% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_funcs.R
\name{per_day}
\alias{per_day}
\title{Judge daily frequency of date/time vars}
\usage{
per_day(datims, p_start, p_end, freq, n_days = 3)
}
\arguments{
\item{datims}{vector of POSIXct date/times.}

\item{p_start}{start of period to be considered, POSIXct date/time.}

\item{p_end}{end of period to be considered, POSIXct date/time.}

\item{freq}{frequency required per 24 hour period.}

\item{n_days}{number of days to be considered, from p_start. Only the
earliest of p_end/p_start  + n_days is taken.}
}
\description{
Gives T/F for whether a list of date/times are at a certain frequency per
day (or part thereof). Requires lubridate:: package.
}
\examples{

datims <- c("2015-12-06 08:44:00",
"2015-12-06 12:00:00", "2015-12-06 13:00:00", "2015-12-06 16:00:00",
"2015-12-06 16:00:00", "2015-12-06 20:20:00", "2015-12-06 23:45:00",
"2015-12-07 04:00:00", "2015-12-07 16:45:00", "2015-12-07 19:45:00",
"2015-12-07 23:30:00", "2015-12-08 04:00:00", "2015-12-08 07:30:00",
"2015-12-08 12:00:00")
datims <- as.POSIXct(datims)

per_day(datims, datims[1], datims[14], freq = 6, n_days = 3)
# FALSE
per_day(datims, datims[1], datims[14], freq = 5, n_days = 3)
# TRUE
}
\keyword{daily}
\keyword{hourly}
\keyword{per}
